Don't use should_process test(?) file
